requirements.txt
librosa==0.9.1
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
pyyaml
mido 
nnAudio==0.3.1 
numpy 
bitstruct 
torchcrepeV2